Python_tracking/requirements.txt

flask 
pyserial 
pandas  
openpyxl 